Apply coding format onto current project

+++ core/src/main/scala/com/intel/sparkColumnarPlugin/expression/CodeGeneration.scala
@@ -25,7 +25,7 @@ object CodeGeneration {
       new ArrowType.Int(64, true)
     case t: FloatType =>
       new ArrowType.FloatingPoint(FloatingPointPrecision.SINGLE)
-    case t: DecimalType =>
+    case t: DoubleType =>
       new ArrowType.FloatingPoint(FloatingPointPrecision.DOUBLE)
     case d: DecimalType =>
       new ArrowType.Decimal(d.precision, d.scale)
